Reconcile to ALGLIB++, except TestC.cpp &Notes.htm ∙	Minor layout changes (e.g. conditionals and arithmetic operations, removing redundant brackets, comment formatting, etc.). ∙	Numerals converted to double (in some places only), excesses (double) typecasts eliminated. ∙	1.0 E n, 1.0 e n converted to numeral form, for n ∈ [-8,+8]; 	1.0 e n converted to 1.0 E n most elsewhere outside of comments and strings. ∙	Normalize: X * Constant ⇒ Constant * X. ∙	Normalize: X -= 1.0 and X += 1.0 to X-- and X++ ∙	In-line arithmetic (e.g. X + 0 ⇒ X, 1 + 2 ⇒ 3, 0 * X ⇒ X, 1 * 2 + 1 ⇒ 3, etc.) ∙	Conversions X <= Y - 1 ⇒ X < Y done in those in places other than loops or branches that were missed in a previous pass. ∙	Brackets (e.g. (*X) around pointer parameters in function bodies) removed. ∙	Relabel the messages in some of the ae_assert calls (for now: just those used with the serialization routines). ∙	TestC.cpp: Synthesis of more randommid() and hqrndmiduniformr() routines, missed on a previous pass. ∙	Integrating other recoding and redocumentation changes already made in AlgLib++.

+++ Ap.cpp
@@ -2796,19 +2796,19 @@ double rboundval(double x, double b1, double b2) {
 }
 
 double randomreal() {
-   const double mx = (double)RAND_MAX + 1.0;
+   const double mx = RAND_MAX + 1.0;
    return (rand() + rand() / mx) / mx;
 }
 
 double randommid() {
-   const double mx = (double)RAND_MAX + 1.0;
+   const double mx = RAND_MAX + 1.0;
    return 2.0 * (rand() + rand() / mx) / mx - 1.0;
 }
 
 ae_int_t randominteger(ae_int_t maxv) { return rand() % maxv; }
 
 bool randombool(double p/* = 0.5*/) {
-   const double mx = (double)RAND_MAX + 1.0;
+   const double mx = RAND_MAX + 1.0;
    return rand() + rand()/mx <= p * mx;
 }
 
@@ -5174,19 +5174,19 @@ double maxreal(double x, double y) { return x > y ? x : y; }
 double sqr(double x) { return x * x; }
 
 double randomreal() {
-   const double mx = (double)RAND_MAX + 1.0;
+   const double mx = RAND_MAX + 1.0;
    return (rand() + rand() / mx) / mx;
 }
 
 double randommid() {
-   const double mx = (double)RAND_MAX + 1.0;
+   const double mx = RAND_MAX + 1.0;
    return 2.0 * (rand() + rand() / mx) / mx - 1.0;
 }
 
 ae_int_t randominteger(ae_int_t maxv) { return (ae_int_t)rand() % maxv; }
 
 bool randombool(double p/* = 0.5*/) {
-   const double mx = (double)RAND_MAX + 1.0;
+   const double mx = RAND_MAX + 1.0;
    return rand() + rand()/mx <= p * mx;
 }
